+++ man-roxygen/getset-visible-rowspec-example.R
@@ -5,4 +5,3 @@
 #' set_<%= attr_name %>(ht, <%= attr_val %>)
 #' set_<%= attr_name %>(ht, 1:2, 1, <%= attr_val %>)
 #' set_<%= attr_name %>(ht, 1:2, 1:2, c(<%= attr_val %>, <%= attr_val2 %>), byrow = TRUE)
-#' set_<%= attr_name %>(ht, where(ht == 1), <%= attr_val %>)